Move ScreenshotManager functionality to ScreenCaptureManager
